:recycle: refactor frame styles and components for better reuse and maintainability #23
