+++ src/components/frame/constants.ts
@@ -1,12 +1,13 @@
-// iphone 14 pro max
-export const MAX_HEIGHT = 431;
-export const MAX_WIDTH = 932;
-export const MAX_FRAME_HEIGHT = `${MAX_HEIGHT}px`;
-export const MAX_FRAME_WIDTH = `${MAX_WIDTH}px`;
+import { vars } from '@/styles/vars.css';
 
-// iphone se
-export const MIN_WIDTH = 667;
-export const MIN_HEIGHT = 320;
-export const PADDING = 16;
-export const MIN_FRAME_HEIGHT = `${MIN_HEIGHT}px`;
-export const MIN_FRAME_HEIGHT_WITH_PADDING = `${MIN_HEIGHT - PADDING * 2}px`;
+export const FRAME_STYLE = {
+  width: {
+    max: '932px', // iphone 14 pro max
+    min: '667px', // iphone se
+  },
+  height: {
+    max: '431px',
+    min: '320px',
+  },
+  padding: vars.space['4x'],
+};
