+++ src/components/frame/index.tsx
@@ -1,13 +1,2 @@
-import { ReactNode } from 'react';
-
-import * as styles from './index.css';
-
-interface Props {
-  children: ReactNode;
-}
-
-const BasicContentFrame = ({ children }: Props) => {
-  return <main className={styles.frame}>{children}</main>;
-};
-
-export default BasicContentFrame;
+export { default } from './mobile-fullscreen';
+export { FRAME_STYLE } from './constants';
